WIP: relocate autorun_extractors/localizers off vtsearch.state
Phase 3 (easy bit) of extract-library.md: the autorun extractor/localizer
registries are app-side policy, not library state, so they move out of
the library-candidate vtsearch/state/ package into a new app-level module
vtsearch/autorun_processors.py.

- New file vtsearch/autorun_processors.py owns the dicts + CRUD with its
  own RLock (independent of _state_lock so the library is fully isolated)
- state/core.py: drop the module-level autorun_extractors/localizers dicts
- state/processors.py: deleted (every function moved)
- state/__init__.py: drop the autorun re-exports and the processors.py
  description from the module docstring
- routes/processors/{crud,scoring}.py: import from the new module
- tests/conftest.py: replace direct dict-clear with clear_all_autorun()
- tests/detectors/{test_processors,test_extractors}.py: update 6 import
  sites that previously reached into vtsearch.state for the autorun names
- CLAUDE.md: drop the state/processors.py mention from the state/ blurb
  and add a one-liner for the new autorun_processors.py module
- docs/plans/extract-library.md: check off this follow-up; resolve the
  cross-boundary note in the "what goes where" section
