+++ CLAUDE.md
@@ -165,7 +165,8 @@ A flow can legitimately carry both: a nested view shows `← Back` at the top to
 - `vtsearch/labels/sync.py` — Labelset source sync utilities: `sync_to_labelset_source()` (auto-export on vote change) and `sync_from_labelset_source()` (manual import), with `_syncing` guard to prevent circular re-export
 - `vtsearch/media/` — Media type plugins: audio, image, text, video, document
 - `vtsearch/converters/` — Media converters: document→image, document→text, video→audio, video→image, audio→image (spectrogram), audio→text (Whisper ASR), image→text (OCR). Each is a `MediaConverter` (a `PluginBase`) that can declare user-configurable params via `fields: list[PluginField]`. `convert(media, params)` reads those at conversion time. Used by the multi-media import flow (see `docs/plans/multi-media-import.md`) and the legacy per-importer `converters` field
-- `vtsearch/state/` — Multi-dataset / multi-detector global state. `core.py` owns `DatasetContext`, `DetectorContext`, `_state_lock`, the `medias`/votes proxy objects, and the context registries. Sub-modules split the operations by concern: `votes.py` (toggle_vote / apply_label / clear_votes), `clicks.py` (vote click times), `processors.py` (autorun extractor/localizer CRUD), `diversity.py` (diversity tree), `media_lookup.py` (origin-keyed lookup, collapse_duplicates). The package `__init__.py` re-exports every public name and hosts a handful of cross-cutting helpers (`snapshot_medias`, `clear_medias`, `clear_all`, `get_inclusion`/`set_inclusion`, dataset-display-name and safe-thresholds accessors)
+- `vtsearch/state/` — Multi-dataset / multi-detector global state. `core.py` owns `DatasetContext`, `DetectorContext`, `_state_lock`, the `medias`/votes proxy objects, and the context registries. Sub-modules split the operations by concern: `votes.py` (toggle_vote / apply_label / clear_votes), `clicks.py` (vote click times), `diversity.py` (diversity tree), `media_lookup.py` (origin-keyed lookup, collapse_duplicates). The package `__init__.py` re-exports every public name and hosts a handful of cross-cutting helpers (`snapshot_medias`, `clear_medias`, `clear_all`, `get_inclusion`/`set_inclusion`, dataset-display-name and safe-thresholds accessors)
+- `vtsearch/autorun_processors.py` — App-side singleton holding the `autorun_extractors` / `autorun_localizers` dicts and their CRUD. Lives outside the library-candidate `state/` package because which processors run on every newly-loaded media is user-facing policy. See `docs/plans/extract-library.md` Phase 3.
 - `vtsearch/plugins/` — `PluginRegistry`, `PluginBase`, `PluginField`, and the sentinel-based auto-discovery scanner shared by every plugin family (dataset importers, exporters, label importers/sources, processor importers, settings importers/exporters/sources)
 - `vtsearch/sync/` — Generic `SyncSource[LoadT, SaveT]` base class (built on `PluginBase`) shared by `SettingsSource` and `LabelsetSource` for bidirectional load/save
 - `vtsearch/concurrency/` — Async job manager (`async_jobs.py` — `AsyncJob`, `JobManager`, `eval_jobs`, `learned_sort_jobs`), memory-aware worker capping (`memory_budget.py` — `cap_workers_by_memory`), and the long-running-operation progress infrastructure (`progress.py` — `ProgressTracker`, `LoadingTasksTracker`, `update_progress`/`get_progress` + sort/eval/find variants, `cancel_dataset_progress`, `check_dataset_cancelled`). The labeling-session analyzer in `vtsearch/detectors/labeling_progress.py` is a separate, unrelated module that caches per-step trained MLPs and computes stopping-condition metrics — the historical name collision with this package was resolved by renaming `models/progress.py` to `models/labeling_progress.py` (since moved into `vtsearch/detectors/`)
