+++ vtsearch/state/core.py
@@ -773,8 +773,3 @@ class with_detector_context:
 
     def __exit__(self, *exc_info: object) -> None:
         set_thread_detector_context(self._previous_ctx)
-
-
-# Autorun extractors/localizers are GLOBAL (not per-dataset).
-autorun_extractors: dict[str, dict[str, Any]] = {}
-autorun_localizers: dict[str, dict[str, Any]] = {}
